+++ .sampo/changesets/958-split-typia-llm.md
@@ -1,0 +1,6 @@
+---
+npm/@wp-typia/project-tools: patch
+npm/wp-typia: patch
+---
+
+Split Typia LLM runtime rendering, sync, projection, and OpenAPI constraint helpers into focused modules.
