+++ packages/wp-typia-project-tools/src/runtime/typia-llm.ts
@@ -1,1168 +1,36 @@
-import { mkdir, readFile, writeFile } from 'node:fs/promises';
-import path from 'node:path';
-
-import type {
-  ArtifactSyncExecutionOptions,
-  EndpointManifestDefinition,
-  EndpointManifestEndpointDefinition,
-} from '@wp-typia/block-runtime/metadata-core';
-import type { ILlmFunction, ILlmSchema, ILlmStructuredOutput } from 'typia';
-import {
-  getOptionalNodeErrorCode,
-  isFileNotFoundError,
-} from './fs-async.js';
-import { cloneJsonValue } from './json-utils.js';
-import {
-  normalizeEndpointAuthDefinition,
-  type EndpointAuthIntent,
-  type EndpointWordPressAuthDefinition,
-  type JsonSchemaObject,
-  type OpenApiDocument,
-  type OpenApiOperation,
-} from './schema-core.js';
-
 /**
- * Method-level descriptor projected from an endpoint manifest for a generated
- * `typia.llm` controller interface.
- */
-export interface TypiaLlmEndpointMethodDescriptor {
-  /** Normalized auth intent derived from the endpoint manifest. */
-  authIntent: EndpointAuthIntent;
-  /** Optional auth mode propagated from legacy manifest metadata. */
-  authMode?: EndpointManifestEndpointDefinition['authMode'];
-  /** Human-readable operation summary. */
-  description?: string;
-  /** Input type used by the generated controller method, or null for no input. */
-  inputTypeName: string | null;
-  /** Type names that must be imported when inputTypeName is an inline composite. */
-  inputTypeImportNames?: readonly string[];
-  /** HTTP method from the source endpoint manifest. */
-  method: EndpointManifestEndpointDefinition['method'];
-  /** Stable operation identifier used as the generated method name. */
-  operationId: string;
-  /** Output type used by the generated controller method. */
-  outputTypeName: string;
-  /** REST path from the source endpoint manifest. */
-  path: string;
-  /** Tags copied from the source endpoint manifest. */
-  tags: readonly string[];
-  /** Optional WordPress auth overlay preserved for downstream adapters. */
-  wordpressAuth?: EndpointWordPressAuthDefinition;
-}
-
-/**
- * Configures the generated TypeScript module that asks `typia.llm` to derive
- * function-calling and structured-output artifacts from canonical contracts.
- */
-export interface RenderTypiaLlmModuleOptions {
-  /** Export name for the generated `typia.llm.application<...>()` value. */
-  applicationExportName: string;
-  /** Generated controller interface name. */
-  interfaceName: string;
-  /** Endpoint manifest that owns REST operation metadata and source type names. */
-  manifest: EndpointManifestDefinition;
-  /** Export name for the generated `typia.llm.structuredOutput<...>()` value. */
-  structuredOutputExportName: string;
-  /** TypeScript output type used for structured-output projection. */
-  structuredOutputTypeName: string;
-  /** Import path from the generated module to the canonical TypeScript contracts. */
-  typesImportPath: string;
-}
-
-/**
- * Configures writing or checking the generated `typia.llm` adapter module.
- */
-export interface SyncTypiaLlmAdapterModuleOptions
-  extends ArtifactSyncExecutionOptions,
-    RenderTypiaLlmModuleOptions {
-  /** Destination path for the generated TypeScript adapter module. */
-  generatedSourceFile: string;
-}
-
-/**
- * Describes the generated adapter module and where it was written or checked.
- */
-export interface SyncTypiaLlmAdapterModuleResult {
-  /** True when the source file was only verified, false when it was written. */
-  check: boolean;
-  /** Generated endpoint method descriptors used by the module renderer. */
-  methodDescriptors: TypiaLlmEndpointMethodDescriptor[];
-  /** Rendered TypeScript module source. */
-  source: string;
-  /** Destination path for the generated TypeScript adapter module. */
-  sourceFile: string;
-}
-
-/**
- * Function schema shape projected from a compiled `typia.llm.application(...)`
- * result.
- */
-export interface TypiaLlmFunctionLike
-  extends Pick<ILlmFunction, 'description' | 'name' | 'output' | 'parameters'> {
-  /** Function description generated by `typia.llm`. */
-  description?: string;
-  /** Function name generated by `typia.llm`. */
-  name: string;
-  /** Optional output schema generated by `typia.llm`. */
-  output?: ILlmSchema.IParameters;
-  /** Function parameter schema generated by `typia.llm`. */
-  parameters: ILlmSchema.IParameters;
-  /** Optional tags generated by `typia.llm`. */
-  tags?: readonly string[];
-}
-
-/**
- * Application shape projected from a compiled `typia.llm.application(...)`
- * result.
- */
-export interface TypiaLlmApplicationLike {
-  /** Function schemas generated by `typia.llm`. */
-  functions: readonly TypiaLlmFunctionLike[];
-}
-
-/**
- * Structured-output shape projected from a compiled
- * `typia.llm.structuredOutput(...)` result.
- */
-export interface TypiaLlmStructuredOutputLike
-  extends Pick<ILlmStructuredOutput, 'parameters'> {
-  /** Structured output parameters generated by `typia.llm`. */
-  parameters: ILlmStructuredOutput['parameters'];
-}
-
-/**
- * JSON-friendly function artifact for downstream adapter consumers.
- */
-export interface ProjectedTypiaLlmFunctionArtifact {
-  /** Function description generated by `typia.llm`. */
-  description?: string;
-  /** Function name generated by `typia.llm`. */
-  name: string;
-  /** Optional output schema generated by `typia.llm`. */
-  output?: ILlmSchema.IParameters;
-  /** Function parameter schema generated by `typia.llm`. */
-  parameters: ILlmSchema.IParameters;
-  /** Optional tags generated by `typia.llm`. */
-  tags?: string[];
-}
-
-/**
- * JSON-friendly application artifact for downstream adapter consumers.
- */
-export interface ProjectedTypiaLlmApplicationArtifact {
-  /** Function-calling artifacts generated by `typia.llm`. */
-  functions: ProjectedTypiaLlmFunctionArtifact[];
-  /** Source metadata that keeps the artifact tied to canonical contracts. */
-  generatedFrom: {
-    baselineOpenApiPath?: string;
-    blockSlug: string;
-    manifestSource: 'endpoint-manifest+typescript';
-  };
-}
-
-/**
- * JSON-friendly structured-output artifact for downstream adapter consumers.
- */
-export interface ProjectedTypiaLlmStructuredOutputArtifact {
-  /** Source metadata that keeps the artifact tied to canonical contracts. */
-  generatedFrom: {
-    aiSchemaPath?: string;
-    blockSlug: string;
-    outputTypeName: string;
-  };
-  /** Structured output parameters generated by `typia.llm`. */
-  parameters: ILlmStructuredOutput['parameters'];
-}
-
-/**
- * Configures optional OpenAPI-backed constraint restoration for projected
- * `typia.llm` function artifacts.
- */
-export interface ProjectTypiaLlmOpenApiProjectionOptions {
-  /** Canonical endpoint-aware OpenAPI document for the projected contracts. */
-  openApiDocument: OpenApiDocument;
-  /** Optional override for resolving one function schema to an operation id. */
-  resolveOperationId?: (
-    functionSchema: TypiaLlmFunctionLike,
-    functionArtifact: ProjectedTypiaLlmFunctionArtifact,
-  ) => string;
-}
-
-/**
- * Configures projection of a compiled `typia.llm.application(...)` result into
- * the JSON-friendly adapter artifact.
- */
-export interface ProjectTypiaLlmApplicationArtifactOptions {
-  /** Compiled application value produced by `typia.llm.application(...)`. */
-  application: TypiaLlmApplicationLike;
-  /** Source metadata for the generated JSON artifact. */
-  generatedFrom: ProjectedTypiaLlmApplicationArtifact['generatedFrom'];
-  /** Optional OpenAPI projection that restores canonical schema constraints. */
-  openApiProjection?: ProjectTypiaLlmOpenApiProjectionOptions;
-  /** Optional hook for adapter-specific function schema normalization. */
-  transformFunction?: (
-    functionArtifact: ProjectedTypiaLlmFunctionArtifact,
-    functionSchema: TypiaLlmFunctionLike,
-  ) => ProjectedTypiaLlmFunctionArtifact;
-}
-
-/**
- * Configures direct OpenAPI-backed constraint restoration for one projected
- * function artifact.
- */
-export interface ApplyOpenApiConstraintsToTypiaLlmFunctionArtifactOptions {
-  /** Projected function artifact to enrich with canonical constraints. */
-  functionArtifact: ProjectedTypiaLlmFunctionArtifact;
-  /** Canonical endpoint-aware OpenAPI document. */
-  openApiDocument: OpenApiDocument;
-  /** Operation id that owns the request/response contract for this function. */
-  operationId: string;
-}
-
-/**
- * Configures projection of a compiled `typia.llm.structuredOutput(...)` result
- * into the JSON-friendly adapter artifact.
- */
-export interface ProjectTypiaLlmStructuredOutputArtifactOptions {
-  /** Source metadata for the generated JSON artifact. */
-  generatedFrom: ProjectedTypiaLlmStructuredOutputArtifact['generatedFrom'];
-  /** Compiled structured-output value produced by `typia.llm.structuredOutput(...)`. */
-  structuredOutput: TypiaLlmStructuredOutputLike;
-}
-
-type GeneratedTypiaLlmArtifactFile = {
-  content: string;
-  filePath: string;
-};
-
-const TYPESCRIPT_RESERVED_WORDS = new Set([
-  'abstract',
-  'any',
-  'as',
-  'asserts',
-  'async',
-  'await',
-  'bigint',
-  'boolean',
-  'break',
-  'case',
-  'catch',
-  'class',
-  'const',
-  'constructor',
-  'continue',
-  'debugger',
-  'declare',
-  'default',
-  'delete',
-  'do',
-  'else',
-  'enum',
-  'export',
-  'extends',
-  'false',
-  'finally',
-  'for',
-  'from',
-  'function',
-  'get',
-  'global',
-  'if',
-  'implements',
-  'import',
-  'in',
-  'infer',
-  'instanceof',
-  'interface',
-  'is',
-  'keyof',
-  'let',
-  'module',
-  'namespace',
-  'never',
-  'new',
-  'null',
-  'number',
-  'object',
-  'of',
-  'package',
-  'private',
-  'protected',
-  'public',
-  'readonly',
-  'require',
-  'return',
-  'satisfies',
-  'set',
-  'static',
-  'string',
-  'super',
-  'switch',
-  'symbol',
-  'this',
-  'throw',
-  'true',
-  'try',
-  'type',
-  'typeof',
-  'undefined',
-  'unique',
-  'unknown',
-  'var',
-  'void',
-  'while',
-  'with',
-  'yield',
-]);
-
-const JSON_SCHEMA_CONSTRAINT_KEYS = [
-  'additionalProperties',
-  'const',
-  'default',
-  'enum',
-  'exclusiveMaximum',
-  'exclusiveMinimum',
-  'format',
-  'maxItems',
-  'maxLength',
-  'maximum',
-  'minItems',
-  'minLength',
-  'minimum',
-  'multipleOf',
-  'pattern',
-  'type',
-] as const;
-
-function cloneJsonValueIfDefined<T>(value: T | undefined): T | undefined {
-  return value === undefined ? undefined : cloneJsonValue(value);
-}
-
-function isJsonSchemaObject(value: unknown): value is JsonSchemaObject {
-  return typeof value === 'object' && value !== null && !Array.isArray(value);
-}
-
-/**
- * Narrow a `typia.llm` artifact field before applying JSON Schema mutations.
+ * Public compatibility facade for Typia LLM runtime helpers.
  *
- * @param value Candidate schema value emitted by `typia.llm`.
- * @param context Human-readable artifact path for diagnostics.
- * @returns The value narrowed to the shared JSON Schema object shape.
+ * Keep `@wp-typia/project-tools/typia-llm` stable while rendering, sync,
+ * projection, and OpenAPI constraint restoration live in focused modules.
  */
-export function assertJsonSchemaObject(
-  value: unknown,
-  context: string,
-): JsonSchemaObject {
-  if (isJsonSchemaObject(value)) {
-    return value;
-  }
-
-  throw new Error(`${context} must be a JSON Schema object.`);
-}
-
-function decodeJsonPointerSegment(segment: string): string {
-  return segment.replace(/~1/g, '/').replace(/~0/g, '~');
-}
-
-function resolveOpenApiReferenceTarget(
-  document: OpenApiDocument,
-  reference: string,
-): unknown {
-  if (!reference.startsWith('#/')) {
-    throw new Error(`Unsupported OpenAPI schema reference "${reference}".`);
-  }
-
-  let current: unknown = document;
-  for (const rawSegment of reference.slice(2).split('/')) {
-    const segment = decodeJsonPointerSegment(rawSegment);
-    if (Array.isArray(current)) {
-      const index = Number.parseInt(segment, 10);
-      current = Number.isInteger(index) ? current[index] : undefined;
-      continue;
-    }
-    if (!isJsonSchemaObject(current)) {
-      current = undefined;
-      break;
-    }
-    current = current[segment];
-  }
-
-  if (!isJsonSchemaObject(current)) {
-    throw new Error(
-      `Unable to resolve OpenAPI schema reference "${reference}".`,
-    );
-  }
-
-  return current;
-}
-
-function resolveOpenApiSchemaObject(
-  document: OpenApiDocument,
-  schema: unknown,
-  seenReferences: ReadonlySet<string> = new Set(),
-): JsonSchemaObject {
-  if (!isJsonSchemaObject(schema)) {
-    return {};
-  }
-
-  const reference = schema.$ref;
-  if (typeof reference !== 'string') {
-    return schema;
-  }
-
-  if (seenReferences.has(reference)) {
-    return schema;
-  }
-
-  return resolveOpenApiSchemaObject(
-    document,
-    resolveOpenApiReferenceTarget(document, reference),
-    new Set([...seenReferences, reference]),
-  );
-}
-
-function collectOpenApiSeenReferences(
-  schema: JsonSchemaObject,
-  seenReferences: ReadonlySet<string>,
-): ReadonlySet<string> {
-  const reference = schema.$ref;
-  if (typeof reference !== 'string' || seenReferences.has(reference)) {
-    return seenReferences;
-  }
-
-  return new Set([...seenReferences, reference]);
-}
-
-/**
- * Copy constraint-oriented JSON Schema keywords from `source` into `target`.
- *
- * This helper intentionally mutates and returns `target` so nested OpenAPI
- * constraint restoration can preserve the projected typia.llm schema shape.
- * Callers that need immutable behavior should clone the target before calling.
- */
-function mergeJsonSchemaConstraintProperties(
-  document: OpenApiDocument,
-  target: JsonSchemaObject,
-  source: JsonSchemaObject,
-  seenReferences: ReadonlySet<string> = new Set(),
-): JsonSchemaObject {
-  const nextSeenReferences = collectOpenApiSeenReferences(
-    source,
-    seenReferences,
-  );
-  const resolvedSource = resolveOpenApiSchemaObject(
-    document,
-    source,
-    seenReferences,
-  );
-  const merged = target;
-
-  for (const key of JSON_SCHEMA_CONSTRAINT_KEYS) {
-    if (resolvedSource[key] !== undefined) {
-      merged[key] = cloneJsonValue(resolvedSource[key]);
-    }
-  }
-
-  if (Array.isArray(resolvedSource.required)) {
-    merged.required = resolvedSource.required.filter(
-      (value): value is string => typeof value === 'string',
-    );
-  }
-
-  if (Array.isArray(resolvedSource.items)) {
-    merged.items = cloneJsonValue(resolvedSource.items) as JsonSchemaObject[];
-  } else if (isJsonSchemaObject(resolvedSource.items)) {
-    const nextItems = isJsonSchemaObject(merged.items) ? merged.items : {};
-    merged.items = mergeJsonSchemaConstraintProperties(
-      document,
-      nextItems,
-      resolvedSource.items,
-      nextSeenReferences,
-    );
-  }
-
-  if (isJsonSchemaObject(resolvedSource.properties)) {
-    const targetProperties = isJsonSchemaObject(merged.properties)
-      ? merged.properties
-      : {};
-
-    for (const [propertyName, propertySchema] of Object.entries(
-      resolvedSource.properties,
-    )) {
-      if (!isJsonSchemaObject(propertySchema)) {
-        continue;
-      }
-
-      const nextProperty = isJsonSchemaObject(targetProperties[propertyName])
-        ? (targetProperties[propertyName] as JsonSchemaObject)
-        : {};
-      targetProperties[propertyName] = mergeJsonSchemaConstraintProperties(
-        document,
-        nextProperty,
-        propertySchema,
-        nextSeenReferences,
-      );
-    }
-
-    merged.properties = targetProperties;
-  }
-
-  return merged;
-}
-
-function findOpenApiOperationById(
-  document: OpenApiDocument,
-  operationId: string,
-): OpenApiOperation | null {
-  for (const pathItem of Object.values(document.paths)) {
-    if (!isJsonSchemaObject(pathItem)) {
-      continue;
-    }
-
-    for (const method of ['delete', 'get', 'patch', 'post', 'put'] as const) {
-      const operation = resolveOpenApiSchemaObject(document, pathItem[method]);
-      if (
-        !isJsonSchemaObject(operation) ||
-        typeof operation.operationId !== 'string'
-      ) {
-        continue;
-      }
-
-      if (operation.operationId === operationId) {
-        return operation as OpenApiOperation;
-      }
-    }
-  }
-
-  return null;
-}
-
-function resolveOpenApiRequestBodySchema(
-  operation: OpenApiOperation,
-  document: OpenApiDocument,
-): JsonSchemaObject | null {
-  const requestBody = resolveOpenApiSchemaObject(document, operation.requestBody);
-  const content = isJsonSchemaObject(requestBody.content)
-    ? requestBody.content
-    : null;
-  const jsonMediaType =
-    content && isJsonSchemaObject(content['application/json'])
-      ? content['application/json']
-      : null;
-  const schema = jsonMediaType?.schema;
-  if (!isJsonSchemaObject(schema)) {
-    return null;
-  }
-
-  return resolveOpenApiSchemaObject(document, schema);
-}
-
-function resolveOpenApiSuccessResponseSchema(
-  operation: OpenApiOperation,
-  document: OpenApiDocument,
-): JsonSchemaObject | null {
-  for (const [statusCode, response] of Object.entries(operation.responses)) {
-    const resolvedResponse = resolveOpenApiSchemaObject(document, response);
-    if (
-      !/^2(?:\d\d|XX)$/u.test(statusCode) ||
-      !isJsonSchemaObject(resolvedResponse)
-    ) {
-      continue;
-    }
-
-    const content = isJsonSchemaObject(resolvedResponse.content)
-      ? resolvedResponse.content
-      : null;
-    const jsonMediaType =
-      content && isJsonSchemaObject(content['application/json'])
-        ? content['application/json']
-        : null;
-    const schema = jsonMediaType?.schema;
-    if (!isJsonSchemaObject(schema)) {
-      continue;
-    }
-
-    return resolveOpenApiSchemaObject(document, schema);
-  }
-
-  return null;
-}
-
-function getOrCreateObjectProperty(
-  target: JsonSchemaObject,
-  propertyName: string,
-): JsonSchemaObject {
-  const targetProperties = isJsonSchemaObject(target.properties)
-    ? target.properties
-    : {};
-  const nextProperty = isJsonSchemaObject(targetProperties[propertyName])
-    ? (targetProperties[propertyName] as JsonSchemaObject)
-    : {};
-
-  targetProperties[propertyName] = nextProperty;
-  target.properties = targetProperties;
-  return nextProperty;
-}
-
-function applyOpenApiQueryParameterConstraints(
-  target: JsonSchemaObject,
-  operation: OpenApiOperation,
-  document: OpenApiDocument,
-): void {
-  for (const parameter of operation.parameters ?? []) {
-    const resolvedParameter = resolveOpenApiSchemaObject(document, parameter);
-    if (
-      !isJsonSchemaObject(resolvedParameter) ||
-      resolvedParameter.in !== 'query' ||
-      typeof resolvedParameter.name !== 'string'
-    ) {
-      continue;
-    }
-
-    const propertyTarget = getOrCreateObjectProperty(
-      target,
-      resolvedParameter.name,
-    );
-    mergeJsonSchemaConstraintProperties(
-      document,
-      propertyTarget,
-      resolveOpenApiSchemaObject(document, resolvedParameter.schema),
-    );
-
-    if (resolvedParameter.required === true) {
-      const required = new Set(
-        Array.isArray(target.required)
-          ? target.required.filter(
-              (value): value is string => typeof value === 'string',
-            )
-          : [],
-      );
-      required.add(resolvedParameter.name);
-      target.required = [...required];
-    }
-  }
-}
-
-interface EndpointInputTypeDescriptor {
-  importTypeNames: string[];
-  signatureTypeName: string | null;
-}
-
-function getContractSourceTypeName(
-  manifest: EndpointManifestDefinition,
-  endpoint: EndpointManifestEndpointDefinition,
-  contractName: string,
-): string {
-  const contract = manifest.contracts[contractName];
-  if (!contract) {
-    throw new Error(
-      `Endpoint "${endpoint.operationId}" references missing input contract "${contractName}".`,
-    );
-  }
-
-  return contract.sourceTypeName;
-}
-
-function getEndpointInputTypeDescriptor(
-  manifest: EndpointManifestDefinition,
-  endpoint: EndpointManifestEndpointDefinition,
-): EndpointInputTypeDescriptor {
-  if (endpoint.bodyContract && endpoint.queryContract) {
-    const bodyTypeName = getContractSourceTypeName(
-      manifest,
-      endpoint,
-      endpoint.bodyContract,
-    );
-    const queryTypeName = getContractSourceTypeName(
-      manifest,
-      endpoint,
-      endpoint.queryContract,
-    );
-
-    return {
-      importTypeNames: [bodyTypeName, queryTypeName],
-      signatureTypeName: `{ body: ${bodyTypeName}; query: ${queryTypeName} }`,
-    };
-  }
-
-  const contractName =
-    endpoint.method === 'GET'
-      ? endpoint.queryContract ?? null
-      : endpoint.bodyContract ?? endpoint.queryContract ?? null;
-
-  if (!contractName) {
-    return {
-      importTypeNames: [],
-      signatureTypeName: null,
-    };
-  }
-
-  const sourceTypeName = getContractSourceTypeName(
-    manifest,
-    endpoint,
-    contractName,
-  );
-
-  return {
-    importTypeNames: [sourceTypeName],
-    signatureTypeName: sourceTypeName,
-  };
-}
-
-function getEndpointOutputTypeName(
-  manifest: EndpointManifestDefinition,
-  endpoint: EndpointManifestEndpointDefinition,
-): string {
-  const contract = manifest.contracts[endpoint.responseContract];
-  if (!contract) {
-    throw new Error(
-      `Endpoint "${endpoint.operationId}" references missing response contract "${endpoint.responseContract}".`,
-    );
-  }
-
-  return contract.sourceTypeName;
-}
-
-function escapeJsDocLine(line: string): string {
-  return line.replace(/\*\//g, '* /');
-}
-
-function normalizeGeneratedArtifactContent(content: string): string {
-  return content.replace(/\r\n?/g, '\n');
-}
-
-function renderDescriptionJsDocLines(
-  method: TypiaLlmEndpointMethodDescriptor,
-): string[] {
-  const description =
-    method.description && method.description.trim().length > 0
-      ? method.description
-      : method.operationId;
-  const descriptionLines = description
-    .replace(/\r\n?/g, '\n')
-    .split('\n')
-    .map((line) => line.trim().replace(/\s+/g, ' '))
-    .filter(Boolean);
-
-  return (descriptionLines.length > 0
-    ? descriptionLines
-    : [method.operationId]
-  ).map((line) => ` * ${escapeJsDocLine(line)}`);
-}
-
-function renderMethodJsDoc(method: TypiaLlmEndpointMethodDescriptor): string {
-  const lines = [
-    '/**',
-    ...renderDescriptionJsDocLines(method),
-    ' *',
-    ` * REST path: ${method.method} ${method.path}`,
-    ` * Auth intent: ${method.authIntent}`,
-    ...(method.wordpressAuth
-      ? [
-          ` * WordPress auth: ${method.wordpressAuth.mechanism}${
-            method.wordpressAuth.publicTokenField
-              ? ` (field: ${method.wordpressAuth.publicTokenField})`
-              : ''
-          }`,
-        ]
-      : []),
-    ...method.tags.map((tag) => ` * @tag ${escapeJsDocLine(tag)}`),
-    ' */',
-  ];
-
-  return lines.join('\n');
-}
-
-function renderMethodSignature(method: TypiaLlmEndpointMethodDescriptor): string {
-  const inputSignature =
-    method.inputTypeName === null ? '' : `input: ${method.inputTypeName}`;
-  const methodName =
-    /^[$A-Z_][$0-9A-Z_]*$/i.test(method.operationId) &&
-    !TYPESCRIPT_RESERVED_WORDS.has(method.operationId)
-    ? method.operationId
-    : JSON.stringify(method.operationId);
-
-  return `${methodName}(${inputSignature}): ${method.outputTypeName};`;
-}
-
-function renderTypiaLlmModuleFromMethodDescriptors(
-  {
-    applicationExportName,
-    interfaceName,
-    structuredOutputExportName,
-    structuredOutputTypeName,
-    typesImportPath,
-  }: Omit<RenderTypiaLlmModuleOptions, 'manifest'>,
-  methods: readonly TypiaLlmEndpointMethodDescriptor[],
-): string {
-  const importedTypeNames = new Set<string>([structuredOutputTypeName]);
-
-  for (const method of methods) {
-    importedTypeNames.add(method.outputTypeName);
-    if (method.inputTypeImportNames) {
-      for (const inputTypeImportName of method.inputTypeImportNames) {
-        importedTypeNames.add(inputTypeImportName);
-      }
-    } else if (method.inputTypeName) {
-      importedTypeNames.add(method.inputTypeName);
-    }
-  }
-
-  const imports = Array.from(importedTypeNames).sort().join(',\n\t');
-  const methodBlocks = methods
-    .map((method) => {
-      const jsDoc = renderMethodJsDoc(method)
-        .split('\n')
-        .map((line) => `\t${line}`)
-        .join('\n');
-
-      return `${jsDoc}\n\t${renderMethodSignature(method)}`;
-    })
-    .join('\n\n');
-
-  return `import typia from "typia";
-import type {
-\t${imports},
-} from "${typesImportPath}";
-
-export interface ${interfaceName} {
-${methodBlocks}
-}
-
-export const ${applicationExportName} =
-\ttypia.llm.application<${interfaceName}>();
-
-export const ${structuredOutputExportName} =
-\ttypia.llm.structuredOutput<${structuredOutputTypeName}>();
-`;
-}
-
-async function reconcileGeneratedTypiaLlmArtifacts(
-  artifacts: readonly GeneratedTypiaLlmArtifactFile[],
-  check: boolean,
-): Promise<void> {
-  if (!check) {
-    for (const artifact of artifacts) {
-      await mkdir(path.dirname(artifact.filePath), {
-        recursive: true,
-      });
-      await writeFile(artifact.filePath, artifact.content, 'utf8');
-    }
-
-    return;
-  }
-
-  const issues: string[] = [];
-
-  for (const artifact of artifacts) {
-    try {
-      const current = normalizeGeneratedArtifactContent(
-        await readFile(artifact.filePath, 'utf8'),
-      );
-      const expected = normalizeGeneratedArtifactContent(artifact.content);
-      if (current !== expected) {
-        issues.push(`- ${artifact.filePath} (stale)`);
-      }
-    } catch (error) {
-      if (isFileNotFoundError(error)) {
-        issues.push(`- ${artifact.filePath} (missing)`);
-        continue;
-      }
-
-      const code = getOptionalNodeErrorCode(error);
-      issues.push(
-        `- ${artifact.filePath} (unreadable: ${error instanceof Error ? error.message : code ?? 'unknown'})`,
-      );
-    }
-  }
-
-  if (issues.length > 0) {
-    throw new Error(
-      `Generated typia.llm artifacts are missing or stale:\n${issues.join('\n')}`,
-    );
-  }
-}
-
-/**
- * Builds the generated controller method descriptors that bridge endpoint
- * manifests to `typia.llm` TypeScript interfaces.
- *
- * @param manifest Endpoint manifest that owns operation and source type names.
- * @returns Ordered method descriptors matching the manifest endpoints.
- * @throws When an endpoint has ambiguous query/body input mapping.
- * @throws When an endpoint references a missing input or output contract.
- */
-export function buildTypiaLlmEndpointMethodDescriptors(
-  manifest: EndpointManifestDefinition,
-): TypiaLlmEndpointMethodDescriptor[] {
-  return manifest.endpoints.map((endpoint) => {
-    const normalizedAuth = normalizeEndpointAuthDefinition(endpoint);
-    const inputTypeDescriptor = getEndpointInputTypeDescriptor(manifest, endpoint);
-    const inputTypeName = inputTypeDescriptor.signatureTypeName;
-    const shouldUseInlineInputImports =
-      inputTypeName !== null &&
-      inputTypeDescriptor.importTypeNames.length > 0 &&
-      inputTypeDescriptor.importTypeNames[0] !== inputTypeName;
-
-    return {
-      authIntent: normalizedAuth.auth,
-      ...(normalizedAuth.authMode ? { authMode: normalizedAuth.authMode } : {}),
-      description: endpoint.summary,
-      inputTypeName,
-      ...(shouldUseInlineInputImports
-        ? { inputTypeImportNames: inputTypeDescriptor.importTypeNames }
-        : {}),
-      method: endpoint.method,
-      operationId: endpoint.operationId,
-      outputTypeName: getEndpointOutputTypeName(manifest, endpoint),
-      path: endpoint.path,
-      tags: endpoint.tags ?? [],
-      ...(normalizedAuth.wordpressAuth
-        ? { wordpressAuth: normalizedAuth.wordpressAuth }
-        : {}),
-    };
-  });
-}
-
-/**
- * Renders a build-time-only TypeScript module that invokes `typia.llm` against
- * canonical manifest-owned contracts.
- *
- * @param options Render options for the generated adapter module.
- * @returns The generated TypeScript source.
- */
-export function renderTypiaLlmModule({
-  applicationExportName,
-  interfaceName,
-  manifest,
-  structuredOutputExportName,
-  structuredOutputTypeName,
-  typesImportPath,
-}: RenderTypiaLlmModuleOptions): string {
-  return renderTypiaLlmModuleFromMethodDescriptors(
-    {
-      applicationExportName,
-      interfaceName,
-      structuredOutputExportName,
-      structuredOutputTypeName,
-      typesImportPath,
-    },
-    buildTypiaLlmEndpointMethodDescriptors(manifest),
-  );
-}
-
-/**
- * Writes or verifies the generated build-time `typia.llm` adapter module. The
- * returned TypeScript source still needs to be compiled by the consuming
- * project's Typia transformer before JSON artifacts can be read from it.
- *
- * @param options Generated module destination plus render options.
- * @returns Rendered source, method descriptors, and the checked/written file path.
- */
-export async function syncTypiaLlmAdapterModule({
-  check = false,
-  generatedSourceFile,
-  ...renderOptions
-}: SyncTypiaLlmAdapterModuleOptions): Promise<SyncTypiaLlmAdapterModuleResult> {
-  const methodDescriptors = buildTypiaLlmEndpointMethodDescriptors(
-    renderOptions.manifest,
-  );
-  const source = renderTypiaLlmModuleFromMethodDescriptors(
-    renderOptions,
-    methodDescriptors,
-  );
-
-  await reconcileGeneratedTypiaLlmArtifacts(
-    [
-      {
-        content: source,
-        filePath: generatedSourceFile,
-      },
-    ],
-    check,
-  );
-
-  return {
-    check,
-    methodDescriptors,
-    source,
-    sourceFile: generatedSourceFile,
-  };
-}
-
-/**
- * Projects one compiled `typia.llm` function schema into a JSON-friendly
- * artifact record.
- *
- * @param functionSchema Function schema from a compiled `typia.llm.application(...)`.
- * @returns JSON-friendly function artifact.
- */
-export function projectTypiaLlmApplicationFunction(
-  functionSchema: TypiaLlmFunctionLike,
-): ProjectedTypiaLlmFunctionArtifact {
-  return {
-    description: functionSchema.description,
-    name: functionSchema.name,
-    output: cloneJsonValueIfDefined(functionSchema.output),
-    parameters: cloneJsonValue(functionSchema.parameters),
-    ...(functionSchema.tags ? { tags: [...functionSchema.tags] } : {}),
-  };
-}
-
-function cloneProjectedTypiaLlmFunctionArtifact(
-  functionArtifact: ProjectedTypiaLlmFunctionArtifact,
-): ProjectedTypiaLlmFunctionArtifact {
-  return {
-    description: functionArtifact.description,
-    name: functionArtifact.name,
-    output: cloneJsonValueIfDefined(functionArtifact.output),
-    parameters: cloneJsonValue(functionArtifact.parameters),
-    ...(functionArtifact.tags ? { tags: [...functionArtifact.tags] } : {}),
-  };
-}
-
-/**
- * Restores canonical request and response constraints from an endpoint-aware
- * OpenAPI document onto one projected `typia.llm` function artifact.
- *
- * Query-only inputs merge into the root parameter object. Mixed body/query
- * inputs merge into the generated `body` and `query` properties.
- *
- * @param options Projected artifact plus the canonical OpenAPI document.
- * @returns A cloned artifact enriched with OpenAPI-backed constraints when available.
- */
-export function applyOpenApiConstraintsToTypiaLlmFunctionArtifact({
-  functionArtifact,
-  openApiDocument,
-  operationId,
-}: ApplyOpenApiConstraintsToTypiaLlmFunctionArtifactOptions): ProjectedTypiaLlmFunctionArtifact {
-  const constrainedArtifact =
-    cloneProjectedTypiaLlmFunctionArtifact(functionArtifact);
-  const operation = findOpenApiOperationById(openApiDocument, operationId);
-  if (!operation) {
-    return constrainedArtifact;
-  }
-
-  const hasQueryParameters = (operation.parameters ?? []).some(
-    (parameter) => {
-      const resolvedParameter = resolveOpenApiSchemaObject(
-        openApiDocument,
-        parameter,
-      );
-      return (
-        isJsonSchemaObject(resolvedParameter) &&
-        resolvedParameter.in === 'query' &&
-        typeof resolvedParameter.name === 'string'
-      );
-    },
-  );
-  const requestBodySchema = resolveOpenApiRequestBodySchema(
-    operation,
-    openApiDocument,
-  );
-  const parameterSchema = assertJsonSchemaObject(
-    constrainedArtifact.parameters,
-    `typia.llm parameters for "${constrainedArtifact.name}"`,
-  );
-
-  if (requestBodySchema) {
-    if (hasQueryParameters) {
-      mergeJsonSchemaConstraintProperties(
-        openApiDocument,
-        getOrCreateObjectProperty(parameterSchema, 'body'),
-        requestBodySchema,
-      );
-    } else {
-      mergeJsonSchemaConstraintProperties(
-        openApiDocument,
-        parameterSchema,
-        requestBodySchema,
-      );
-    }
-  }
-
-  if (hasQueryParameters) {
-    applyOpenApiQueryParameterConstraints(
-      requestBodySchema
-        ? getOrCreateObjectProperty(parameterSchema, 'query')
-        : parameterSchema,
-      operation,
-      openApiDocument,
-    );
-  }
-
-  if (constrainedArtifact.output) {
-    const outputSchema = assertJsonSchemaObject(
-      constrainedArtifact.output,
-      `typia.llm output for "${constrainedArtifact.name}"`,
-    );
-    const responseSchema = resolveOpenApiSuccessResponseSchema(
-      operation,
-      openApiDocument,
-    );
-    if (responseSchema) {
-      mergeJsonSchemaConstraintProperties(
-        openApiDocument,
-        outputSchema,
-        responseSchema,
-      );
-    }
-  }
-
-  return constrainedArtifact;
-}
-
-/**
- * Projects a compiled `typia.llm.application(...)` result into a JSON-friendly
- * downstream adapter artifact.
- *
- * @param options Compiled application value and source metadata.
- * @returns JSON-friendly application artifact.
- */
-export function projectTypiaLlmApplicationArtifact({
-  application,
-  generatedFrom,
-  openApiProjection,
-  transformFunction,
-}: ProjectTypiaLlmApplicationArtifactOptions): ProjectedTypiaLlmApplicationArtifact {
-  return {
-    functions: application.functions.map((functionSchema) => {
-      const functionArtifact =
-        projectTypiaLlmApplicationFunction(functionSchema);
-      const openApiAlignedArtifact = openApiProjection
-        ? applyOpenApiConstraintsToTypiaLlmFunctionArtifact({
-            functionArtifact,
-            openApiDocument: openApiProjection.openApiDocument,
-            operationId:
-              openApiProjection.resolveOperationId?.(
-                functionSchema,
-                functionArtifact,
-              ) ?? functionSchema.name,
-          })
-        : functionArtifact;
-      const transformedArtifact = transformFunction
-        ? transformFunction(openApiAlignedArtifact, functionSchema)
-        : openApiAlignedArtifact;
-
-      return cloneProjectedTypiaLlmFunctionArtifact(transformedArtifact);
-    }),
-    generatedFrom: cloneJsonValue(generatedFrom),
-  };
-}
-
-/**
- * Projects a compiled `typia.llm.structuredOutput(...)` result into a
- * JSON-friendly downstream adapter artifact.
- *
- * @param options Compiled structured-output value and source metadata.
- * @returns JSON-friendly structured-output artifact.
- */
-export function projectTypiaLlmStructuredOutputArtifact({
-  generatedFrom,
-  structuredOutput,
-}: ProjectTypiaLlmStructuredOutputArtifactOptions): ProjectedTypiaLlmStructuredOutputArtifact {
-  return {
-    generatedFrom: cloneJsonValue(generatedFrom),
-    parameters: cloneJsonValue(structuredOutput.parameters),
-  };
-}
+export type {
+  ApplyOpenApiConstraintsToTypiaLlmFunctionArtifactOptions,
+  ProjectedTypiaLlmApplicationArtifact,
+  ProjectedTypiaLlmFunctionArtifact,
+  ProjectedTypiaLlmStructuredOutputArtifact,
+  ProjectTypiaLlmApplicationArtifactOptions,
+  ProjectTypiaLlmOpenApiProjectionOptions,
+  ProjectTypiaLlmStructuredOutputArtifactOptions,
+  RenderTypiaLlmModuleOptions,
+  SyncTypiaLlmAdapterModuleOptions,
+  SyncTypiaLlmAdapterModuleResult,
+  TypiaLlmApplicationLike,
+  TypiaLlmEndpointMethodDescriptor,
+  TypiaLlmFunctionLike,
+  TypiaLlmStructuredOutputLike,
+} from './typia-llm-types.js';
+export { assertJsonSchemaObject } from './typia-llm-json-schema.js';
+export {
+  applyOpenApiConstraintsToTypiaLlmFunctionArtifact,
+} from './typia-llm-openapi-constraints.js';
+export {
+  projectTypiaLlmApplicationArtifact,
+  projectTypiaLlmApplicationFunction,
+  projectTypiaLlmStructuredOutputArtifact,
+} from './typia-llm-projection.js';
+export {
+  buildTypiaLlmEndpointMethodDescriptors,
+  renderTypiaLlmModule,
+} from './typia-llm-render.js';
+export { syncTypiaLlmAdapterModule } from './typia-llm-sync.js';
